refactor(internal/sidekick/gcloud): split generate.go by topic
The file had grown long and was mixing several jobs. Move the CLI
model types and the code that builds the model into model.go. Move
the Go client package lookup into client.go. Move the resource-path
helpers into path.go. What is left in generate.go is the top-level
flow: Generate, renderMain, writeMain, renderReadme.

This change moves code only. The behavior is unchanged.
